Refactor eagle3
